+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/core/CliktCommand.kt
@@ -3,6 +3,7 @@ package com.github.ajalt.clikt.core
 import com.github.ajalt.clikt.completion.CompletionGenerator
 import com.github.ajalt.clikt.mpp.exitProcessMpp
 import com.github.ajalt.clikt.mpp.readEnvvar
+import com.github.ajalt.clikt.output.CliktConsole
 import com.github.ajalt.clikt.output.HelpFormatter.ParameterHelp
 import com.github.ajalt.clikt.output.TermUi
 import com.github.ajalt.clikt.parameters.arguments.Argument
@@ -243,8 +244,17 @@ abstract class CliktCommand(
      */
     open fun aliases(): Map<String, List<String>> = emptyMap()
 
-    /** Prints [lineSeparator] to `stdout`, or to `stderr` if [err] is true */
-    protected fun echo(err: Boolean = false, lineSeparator: String = currentContext.console.lineSeparator) {
+    /** Print the default [line separator][CliktConsole.lineSeparator] to `stdout` */
+    protected fun echo() {
+        echo("")
+    }
+
+    @Deprecated(
+            message="Specify message explicitly with `err` or `lineSeparator`",
+            replaceWith = ReplaceWith("echo(\"\", err=err, lineSeparator=lineSeparator)")
+    )
+    /** @suppress */
+    protected fun echo(err: Boolean, lineSeparator: String ) {
         echo("", err = err, lineSeparator = lineSeparator)
     }